Create a projects module and move all projects into it
