+++ client/src/app/project-summary/project-summary.component.ts
@@ -2,8 +2,8 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Observable} from 'rxjs';
 import {Account} from '../core/account';
 import {MetadataFormConfig} from '../metadata-schema-form/models/metadata-form-config';
-import getLayout from '../project-create-edit/components/project-metadata-form/layout';
-import * as ingestSchema from '../project-create-edit/schemas/project-ingest-schema.json';
+import getLayout from '../projects/components/project-metadata-form/layout';
+import * as ingestSchema from '../projects/schemas/project-ingest-schema.json';
 import {Project} from '../shared/models/project';
 import {AlertService} from '../shared/services/alert.service';
 import {IngestService} from '../shared/services/ingest.service';
