+++ client/src/app/projects/components/publication-field-group/publication-field-group.component.ts
