+++ client/src/app/projects/components/wrangler-list-input/wrangler-list-input.component.html
