+++ client/src/app/projects/pages/all-projects/all-projects.component.html
