+++ client/src/app/projects/pages/all-projects/all-projects.component.ts
@@ -2,12 +2,12 @@ import {Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {MatPaginator} from '@angular/material/paginator';
 import {Observable} from 'rxjs';
 import {map} from 'rxjs/operators';
-import {Account} from '../core/account';
-import ingestSchema from '../project-create-edit/schemas/project-ingest-schema.json';
-import {ProjectDataSource} from '../shared/data-sources/project-data-source';
-import {PagedData} from '../shared/models/page';
-import {Project, ProjectColumn} from '../shared/models/project';
-import {IngestService} from '../shared/services/ingest.service';
+import {Account} from '../../../core/account';
+import ingestSchema from '../../schemas/project-ingest-schema.json';
+import {ProjectDataSource} from '../../../shared/data-sources/project-data-source';
+import {PagedData} from '../../../shared/models/page';
+import {Project, ProjectColumn} from '../../../shared/models/project';
+import {IngestService} from '../../../shared/services/ingest.service';
 import {MatSelectChange} from '@angular/material/select';
 
 const THIRTY_SECONDS = 30000;
